==> DATA ANALYSIS/02_Inicjowanie_bazy_danych.ipynb <==
 # Opis notatnika
 W tym notatniku skupimy się na stworzeniu dedykowanej bazy danych wraz ze strukturą tabel. Jednak w tym momencie nie będziemy jej uzupełniać danymi - tym zajmiemy się w kolejnej części.
 Plik wraz ze strukturą bazy znajdziesz w folderze `sql/database_schema.sql`.

 > W tym zadaniu konieczne jest posiadanie zainstalowanego `Postgres` na swoim komputrze. Jeżeli z jakichś powodów został on usunięty po kursie `Python - Analiza Danych`, należy ponownie dokonać instalacji.

 > Posiadając działającą instancję serwera, można stworzyć po prostu nową bazę np. `airlines`.

 W trakcie rozwiązywania tego notatnika przydadzą się następujące materiały:
 - `Python - analiza danych -> Przygotowanie do kursu -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych`
 - `Python - analiza danych -> Moduł 1 -> Dzień 3 - PostgreSQL -> SQL i Python`
 - `Python - analiza danych -> Prework -> Podstawy SQL - Praca samodzielna -> Instalacja bazy danych -> Przygotowanie bazy danych`

 Tutaj zaimportuj potrzebne biblioteki

import psycopg2

 ## Połączenie z bazą danych
 Tutaj uzupełnij konfigurację połączenia

connection = psycopg2.connect(host = 'localhost', user = 'postgres', password = 'Kurs', dbname = 'airlines')

Tutaj zdefiniuj zmienną `con` oraz `cursor`

cursor = connection.cursor()

 ## Wczytanie pliku `database_schema.sql`
 Z katalogu `sql` wczytaj plik `database_schema.sql`

sql_file_data = open(r'../sql/database_schema.sql', mode = 'r')

 W tym miejscu odpowiednio rozdziel zawartość pliku `database_schema.sql` na mniejsze kwerendy używając `;`

sql_file_data = sql_file_data.read().replace('\n', '').split(';')

 W tym miejscu wykonaj każdą z kwerend, aby zainicjować strukturę bazy danych

sql_queries = []
for sql_query in sql_file_data:
    sql_queries.append(f'{sql_query};')

 Zatwierdznie wszystkich operacji na bazie, czyli stworzenie tabel

for sql_query in sql_queries:
    cursor.execute(sql_query)

 ### Sprawdzenie
 Uruchom kod poniżej, aby sprawdzić, czy ta część została poprawnie wykonana



def check_if_table_exists(table_name):
    msg = f"Sprawdzam czy istnieje tabela {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    # jeżeli tabela nie istnieje, ten krok zwróci wyjątek
    cursor.execute(query)
    print('OK!')


tables_to_test = [
    'aircraft',
    'airport_weather',
    'flight',
    'airport_list'
]

for table in tables_to_test:
    check_if_table_exists(table)

connection.close()
msg = "Wszystko wygląda OK :) Możesz przejść do kolejnego zadania."
print(msg)

 # Podsumowanie
 Za pomocą tego notatnika została zbudowana w sposób automatyczny nasza docelowa baza danych. Dzięki temu nie musimy się już martwić o jej ręczną przebudowę - w ramach potrzeby wystarczy włączyć notatnik.
 Wykonując kolejny notatnik, sprawimy, że w tabelach pojawią się również dane potrzebne do przeprowadzenia późniejszej analizy oraz wykorzystywane w raportowaniu.
